requirements.txt
asgiref==3.7.2
azure-common==1.1.28
azure-core==1.29.7
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.19.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.1
django-allauth==0.60.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-extra-fields==3.7.0
filelock==3.13.1
filetype==1.2.0
idna==3.6
iniconfig==2.0.0
isodate==0.6.1
msal==1.26.0
msal-extensions==1.1.0
mysql-connector==2.2.9
numpy==1.26.3
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.4.0
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
pytest==8.0.0
pytest-django==4.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3.post1
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
virtualenv==20.25.0
wheel==0.42.0
